Add the driver of drivers.
The driver of drivers is used to gather drivers for, e.g., the console.
This is so we can have multiple output devices. For coreboot, for example,
we have memory and serial consoles. The machinery in coreboot is very complex.
As you can see here, it is insanely simple.

It currently only works for init, write, and close functions; for read it always
returns 0.

Signed-off-by: Ronald G. Minnich <rminnich@gmail.com>