WIP: ast25x0
This is the assembly code from u-bmc, translated to Rust. Having done
this dram training port thing a number of times, I've found it best to
do a reasonably literal translation to HLL, get that working, then
evolve to decent code.

The src/ needs some suggestions and help.

trans.go is a very crude program that can translate simple ARM
assembly to .rs. It's not very capable, but then we don't want it to
be; it's a discardable tool.

Not all the .rs in src/ are used just yet.

I need to talk to bluecmd about values of config variables;
for now they're zero.

Signed-off-by: Ronald G. Minnich <rminnich@gmail.com>