hifive: first commit
There is no soc, no arch, no uart.

But what's nice is when you do a
RUST_TARGET_PATH=$(PWD) cargo make -p release
the right stuff happens in terms of fetching a toolchain.
So consider this a down payment on what we need to do friday.

I'd suggest we just shoot for hardware. coreboot gives us all the info
we need and hardware is just more fun.

Signed-off-by: Ronald G. Minnich <rminnich@gmail.com>