Add the uart to main
There's a problem: any writes to the uart are trashing
the stack and returns from uart writes go off into a
bad place.

When all uart functions are commented out, we get
to the end of main.

Also, I cleaned up the creation of oreboot.bin,
we're using the ELF for qemu now, and I dump
the assembly to make it easier to see what
is going on.

I'm leaving a loop{} in sifive init so we can debug it.

Signed-off-by: Ronald G. Minnich <rminnich@gmail.com>