sifive/qemu.diff: Update for mainline QEMU
Upstream QEMU has updated and the old diff no longer applied. While we
are updating the diff let's just update it to include the latest patches
on the QEMU list adding support for everything oreboot needs.

We need to also update the run Makefile to use the new QEMU arguments.

Once the patches are in upstream QEMU we can remove all the
modifications to QEMU in the testing and documentation.

Signed-off-by: Alistair Francis <alistair.francis@wdc.com>