azure-pipelines.yml: Use a new QEMU branch
The current diff does not apply to the latest master QEMU. Instead of
updating the diff let's base the testing on my branch which is being
sent to upstream QEMU.

Once the patches are in upstream QEMU we can remove all the
modifications to QEMU in the testing and documentation. For now let's
leave everything else as is and do the bare minimum to get the tests
running.

Signed-off-by: Alistair Francis <alistair.francis@wdc.com>