riscv: bootblock: Spin all harts other then 1
On the HiFive Unleashed Hart 0 is a small embedded E core. Let's not run
on that but instead run on Hart 1 which is a Linux capable U Hart. This
ensures that when we jump to Linux we are running on a U core.

When we support SMP we can go back to running oreboot on the E (Hart 0)
and run Linux on the U cores (Hart 1 - 4).

Signed-off-by: Alistair Francis <alistair.francis@wdc.com>