fu540 init: add a hack for broken rust riscv runtime
For now, rust runtime is broken as it demands 16 bit atomic
load and store ops, which riscv can not do.

Add a gross hack to get us building again until they fix their
code.

This allows me to run again.

Arguably, this belongs in arch, but since it should not exist
at all, and I don't want to add more .S files, it goes here.

We'll revert it very soon.

Signed-off-by: Ronald G. Minnich <rminnich@gmail.com>