opentitan: remove architecture reference from mainboard
we found in coreboot that maintaining a specific hierarchy of reference
made life best.

Currently we reference the architecture from both the mainboard and the
cpu. That gets messy, we have learned.

Reference the architecture from the cpu. Implement halt for the
cpu and the soc.

Use soc::halt in the mainboard code.

Signed-off-by: Ronald G. Minnich <rminnich@gmail.com>