WIP
we're going to not bother with the serde, it makes no sense for the need
we have -- convert 5 things to 5 things.

Long term, we need to change cbfs to little endian. Big-endian lost.
The war is OVER.

But for now we'll just do an inline convert, dirty but simple.

Next week.

Signed-off-by: Ronald G. Minnich <rminnich@gmail.com>