Extract method reading device tree header.
It also changes the logic to first read the whole header into 40 byte
array and then convert them to integers. This has the advantage that we
check the Result from pread just once (instead 10 times), and we can
also check the actual size returned by pread (before the code did not
check that we actually read 4 bytes).

Signed-off-by: Tomasz Zurkowski <zurkowski@google.com>