Store only last part of Path in device_tree crate
This allows us to remove MAX_DEPTH and avoid an expensive copy of 1kb
array for every node and property. If someone needs the full path (which
no-one needs now), the client could store it and write each part of the
path locally (which will be more efficient than existing copies of the
whole buffer).

Signed-off-by: Tomasz Zurkowski <zurkowski@google.com>