Add more tests for device tree reader.
While writing tests, I found that there is a bug in SliceReader (it
always returns the size of the input buffer instead of the number of
bytes actually read).

Closes #156 - existing tests cover various cases, including nested
nodes, different type of properties (empty, number, string) and few
error conditions.

Closes #155 - added test cases for empty properties and shows it works
(though because of Driver API, when property is empty the Driver returns
EOF error, but that is WAI).

Signed-off-by: Tomasz Zurkowski <zurkowski@google.com>