Refactor FdtIterator to use streaming iterator API.
Closes #74.

We are dropping the core::Iterator trait from `FdtIterator` as we never
actually benefited from implementing it. Implementing next with custom
lifetimes allows us to make it more performant (`Entry` can reference
internal buffers in `FdtIterator`).

I am also adding error handling to the `FdtIterator` following
`std::io::Lines` approach. I also removed the use of unsafe from device
tree crate (we check if string is indeed utf8 and return error if it is
not).

I also slightly changed the API of Entry to include both StartNode and
EndNode. This is following similar pattern to StAX XML parsers and
allows users of the iterator to write more composable code.

Signed-off-by: Tomasz Zurkowski <zurkowski@google.com>