Add Driver::pread_exact method.
This is similar to std::io::Read::read_exact, that reads the exact
number of bytes to fill in the input buffers. The current code often
assumes that the buffer will be fully filled in (especially when reading
u32 into a 4 byte buffer), this method will make it much easier to write
it correctly.

Signed-off-by: Tomasz Zurkowski <zurkowski@google.com>