make soc a part
This is prior to adding all the little MPx parts to the system.

They are properly part of the soc.

As a first pass, set up a soc struct and set up its new and init
methods.

Longer term, the cpu_init ought to be called from the soc, but one thing
at a time.

Signed-off-by: Ronald G. Minnich <rminnich@gmail.com>