Add ctl and status out-of-band (OOB) operators
Drivers need both IO operators (read/write) and ctl/status
operators. ctl/status operators provide
what is called out-of-band (OOB) operations.

This design provides OOB operations, based on what we've
learned from user mode APIs, kernels, and coreboot.

For progrmas, Plan 9 provides a control channel in the name space, e.g.,
/dev/uart and /dev/uartctl. Unix provides OOB via system calls
such as soctl, ioctl, mctl, and so on. Inside kernels and coreboot, OOB
is provided via the classic struct-full-of-function pointers,
which only grow in size over time.

All these approaches have pluses and minuses. ioctl is nice
in that it provides a read-modify-write OOB operation, and a
decoupling of the call from the opcodes: opcodes can grow
and change, the call signature remains the same. The limits
of ioctl can be seen in that Unix had to later add soctl, mincore,
mctl, etc: the fixed ioctl API did not quite work.

The table of functions provides enumeration of OOB
ops, but is harder and harder to change as time goes by.
The question of how to indicate an operation is not supported
has also never been settled: nil pointer or functions returning
no error? Both are used in Plan 9.

The alternate OOB device (uart/uartctl) is consistent, and works
well in distributed environments, but the need to then have
two full devices is a headache. Further, Plan 9 use of strings
for operations (which we copied initially) is sensible in that world,
but makes no sense in oreboot: the correctness of a string
is not known at build time. In oreboot, we need to know at compile time
if an unknown opcode is being used. Waiting until boot time is
unacceptable.

Because oreboot has a union device, any design needs to take that
into account. Finally, as we found with this simple change, adding
new driver ops is not something we want to do often.

So:
o oreboot will add two new device operators, ctl and status
o ctl performs OOB operations on a device, as define by Rust enums
  o because of the union device, devices will ignore ctl operations
    they do not implement
    e.g. setting baud rate on port 80 console
    e.g. turning off SPI console
  o if a device DOES implement an operation, if that operation
    goes badly, device can return an error
o status returns status as a string
o a function is provided which takes a variadic set of ctl operations
  so that, e.g. ctl(Off, Baud<115200>, On);
  works.
  this lets us avoid, e.g., UART glitching on baud rate change,
  which is often seen in coreboot

The enums are defined in src/lib/consts.

Builds on hifive and rome.

Signed-off-by: Ronald G. Minnich <rminnich@gmail.com>