Refactor FSP code to build for different architectures
Split up FSP and UEFI code so that non arch-dependant code can be put in
the uefi and fsp-common crates and architecture specific code can be put
in fsp-qemu32-sys and fsp-qemu-sys for Ia32 and X64 respectively.

Note that FSP is 32-bit for all platforms except the x64 QEMU build.

Signed-off-by: Jordan Hand <jhand@google.com>