Run all the makefiles for cibuild
Rather than have the top level list of which boards build and which do not, put
a cibuild target in each makefile. It is then easy to turn it on and off in the makefile;
further, the cibuild targets can be individually tested before pushing boards.

If this model works, then we can have simple toplevel targets for mainboards,
and put the logic for a given board in that board's makefile.

This will avoid all the confusion around exceptions that we have today, as well as
supporting a general principle: all make targets should work at top level
and in each mainboard directory.

Signed-off-by: Ronald G. Minnich <rminnich@gmail.com>