Continue to build on the new makefile layout
In the early days, we had no Makefiles; our build strategy revolved
around Cargo.toml and tools like xargo. This did not end well, and
we went with Makefiles.

Our first pass at Makefiles involved a Makefile at the top with
Makefiles only for mainboards. The Makefile at the top had targets
not available in most places, even in mainboards, and that has proven
very inconvenient.

This commit instantiates a new set of rules for how to use make in
oreboot. The goal is to avoid the problems we saw in previous versions
of both oreboot and coreboot.

The rules are simple:
0. There's only one command you need to remember: make
1. any directory with a Cargo.toml should have a Makefile
2. every Makefile should have a common set of rules for
   format
   clippy
   build
   default
3. When possible, these rules should be defined in a
   top level Makefile.inc
4. mainboards have an extended Makefile designed to
   supporting building a target
5. At two level, there are two lists:
   ALLMAKEFILES
   MAINBOARDMAKEFILES
6. For all directories containing a Makefile,
   it shall be possible to run make with the same
   target as at top level
   e.g, make format at top level should work;
   make format should work for any directory with
   a Makefile
7. No default rules that "happen to work" because
   they come first. Example: the $(IMAGE) rule fires
   in the mainboard today because it is first. It
   has no name you can type. So I added a rule:
   mainboard: $(IMAGE)
   so that people know a mainboard target always works.
   That rule is still first, but n.b.: that rule works
   whether it's first or not.

This is not quite complete. But a few more steps should get us there.

Signed-off-by: Ronald G. Minnich <rminnich@gmail.com>