Rename rust-toolchain{,.toml} to prevent old rustup versions from breaking.
The format of rust-toolchain changed at some point to be full toml, but old versions of rustup choke on the new format. Newer versions will additionally accept rust-toolchain.toml, while these files are invisible to old versions. We want to keep old rustup versions working so that a user can "rustup update" and get to a compatible version. Yes, they could cd out of this repo to update, but this error message is confusing and frustrating.

Signed-off-by: Nickolas Fotopoulos <foton@google.com>