sunxi/nezha: add RustSBI for booting into Linux
This changes sections in the dtfs for Linux and the DTB to hand
over to it. The DTB must be after Linux; otherwise, it would fault.
Leave a TODO for bumping rustsbi and using upstream, since it's
currently a fork due to compile issues with using `asm`.
Also update the dependencies.

Signed-off-by: Daniel Maslowski <info@orangecms.org>