Upgrade Rust to nightly-2022-01-11
This renames our custom x86 target to x86_64-unknown-none-oreboot,
because Rust now has it upstream as well, defaulting to lld and with
different linker options which do not work for all of our boards.
For some boards, explicitly set `linker-flavor=gcc` thusly, and for
others, use the renamed custom target.

The macros asm and global_asm are now stabilized, and asm_const is
an unstable feature, so a bunch of files are migrated and we can use
the sbi crate from crates.io.

Bump a few dependencies.

Signed-off-by: Daniel Maslowski <info@orangecms.org>