Drop all mainboards save emulation and sunxi/nezha
We are moving to the embedded hal and other common crates.
This entails dropping our internall driver crates; we've found
that a driver model is not that crucial.

As a first step, remove mainboards that still use the old model.

This being git, they are always there, and we can bring them back
when we are ready.

Signed-off-by: Ronald G. Minnich <rminnich@gmail.com>