bootstrap starfive/visionfive2 up to DRAM init
Note that the DRAM size is hardcoded to 4G for the time being.
The code has been translated from StarFive's U-Boot fork.
Documentation and comments stem largely from the VisionFive 1.
The DDR PHY setup is reworked with simpler, shorter constant arrays.
We will eventually be able to share code between the two SoCs.

A significant change is in the linker script:
The SRAM base address differs.

Some remaining issues:

A test payload is included, though no xtask setup yet to run it easily.
By default, we resume all harts and run U-Boot from flash.
That hangs due to missing bits it expcts, but shows to generally work.

The frequency change in the UART init code is commented out for now
due to clock setup issues. We rely on what the mask ROM did.

This only runs for the UART boot mode as of now. Flashing to be done.

Signed-off-by: Daniel Maslowski <info@orangecms.org>