lib/log: mutex and debug features, drop debug fn
Dan Cross is right; the debug function does not need to come from here.
If one wanted, one could optionally add that to a specific
implementation.

Instead, add some handy functions for debugging here, to print values as
hex, pointers, etc.. They will help you e.g. when the memory map is
unclear and the print! macro would want to use absolute instead of
relative memory addresses. It helped me figure out the JH7110 SoC.

And because some platforms may not allow for using atomics in SRAM, make
the mutex in this crate opt-in.

Signed-off-by: Daniel Maslowski <info@orangecms.org>