Add fdt library
This will be used by xtask to construct images

add test to verify we read in the dtb correctly

fdt: finish up testing; verify output

We can now create and verify the output.

add a find_fdt function, and tests for it

This will find the fdt in the firmware block.
We require it be 0x1000-byte aligned.

turn off layoutflash run in Makefile.mainboard.inc

it is deprecated for using the layoutflash crate

break layoutflash into core and std parts

Signed-off-by: Ronald G Minnich <rminnich@gmail.com>