[qemu-riscv] replace link.ld with one from link-nezha-bt0.ld
This Ld script fails to work once the code starts improting symbols.
problems is we selectively keep some symbols and selectively get rid of
others and it feels arbitrary. I just replaced the whole meat of the
while with structure equivalent to that of link-nezha-bt0.ld.

The errors are generated arbitrarily and it is very hard to give a
solidc case where errors can be reproduced.

Signed-off-by: Ahmed <ahmedsoliman@oddcoder.com>