milk-v/meles: dram: WIP lpddr4x dram training
dram training data is added along with a few helper functions
output of phase1 training matches to the uboot data.
phase2 training started but the output is 0 when i try to
print some of the values from the registers. the uboot code
prints some data corresponding to those registers but oreboot
prints nothing just 0 for all.

Signed-off-by: Kanak Shilledar <kanakshilledar111@protonmail.com>