xtask: rename dist_dir() -> target_dir()
The _target_ directory is not where binaries for distribution land.
Some platforms have processing units with different ISAs, or in Rust
terms, targets.
Hence, distributable/flashable binaries will eventually end up in
the respective platform_dir(). The oreboot build system is on a higher
level than the Rust/cargo build system, not to be mixed up.

Signed-off-by: Daniel Maslowski <info@orangecms.org>